requirements.txt
colorama
requests
openai
python-dotenv